+++ examples/tracer.py
@@ -2,16 +2,14 @@
 import os
 
 from opentelemetry import trace
-from opentelemetry.ext.lightstep import LightStepSpanExporter
+from opentelemetry.ext.lightstep import LightstepSpanExporter
 from opentelemetry.sdk.trace import TracerProvider
 from opentelemetry.sdk.trace.export import BatchExportSpanProcessor
 
-
 # configure the LightStepSpanExporter as our exporter
-exporter = LightStepSpanExporter(
+exporter = LightstepSpanExporter(
     name=os.getenv("LIGHTSTEP_SERVICE_NAME", "test-service-name"),
     token=os.getenv("LIGHTSTEP_ACCESS_TOKEN", ""),
-    verbosity=5,
     host="ingest.staging.lightstep.com",
     service_version=os.getenv("LIGHTSTEP_SERVICE_VERSION", "0.0.1"),
 )